requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Werkzeug>=2.3.3
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.0
pymysql==1.1.0
cryptography>=41.0.0
